WIP test-configs.yaml: fix renaming of Renesas boards
The automatically generated dtb name follows the name of the device
type.  So this doesn't work when using an alias between multiple
device types that use the same dtb.  Explicitly provide the dtb name
instead.

Also, this still doesn't work across kernel revisions as the dtb name
is different for older ones.  Create variants of the device types for
before and after the kernel revision when the dtb names changed.

Fixes: 0d03b7d8b8d4 ("test-configs: handle rename of renesas boards")
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>