build-configs: Enable coverage of 64K pages on arm64
arm64 supports a range of page sizes, including 64K. In testing on other
systems we seem to turn up a small but interesting stream of bugs on 64K
pages which it'd be good to keep an eye on so let's enable a build with
that config option enabled.

This is added for clang-9 simply because we're not really focused on which
compiler is used for the coverage but clang-9 is the only compiler which
has a list of configs already specified so it's the least invasive change.

Signed-off-by: Mark Brown <broonie@kernel.org>