kselftest: use kernelci/test-definitions repository
Use the kernelci/test-definitions repository with a kernelci.org
branch rather than the Linaro master branch so we can control when the
test definitions get updated in production for kernelci.org.  This
repository can also be used for staging.kernelci.org like the other
GitHub kernelci repositories to test changes in test-definitions.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>