Add crypto test plan
This patch adds the first steps to test crypto in kernelCI
First a crypto fragment is added for enabling Linux crypto selftest.
Then a crypto test plan is added.
This test plan is enabled on various device with hardware crypto accelerator.
This test plan is enabled i both lab-baylibre and lab-clabbe where
thoses devices are present.