build-configs.yaml: add riscv to the clang configs
Add the riscv architecture to the list of configs built with clang.
This is used with clang-10 on mainline and clang-11 on linux-next.

Also add the required make options for it, with LLVM_IAS and the GNU
linker rather than lld, and disable CONFIG_EFI since it currently
breaks the build on mainline and linux-next.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>